Rename module, add minDistance and add playerTp logging
