--- src/main/java/me/wallhacks/spark/systems/module/modules/exploit/TpCoordsLogger.java
@@ -1,34 +1,0 @@
-package me.wallhacks.spark.systems.module.modules.exploit;
-
-import java.util.UUID;
-
-import me.wallhacks.spark.Spark;
-import me.wallhacks.spark.event.player.PacketReceiveEvent;
-import me.wallhacks.spark.systems.module.Module;
-import me.wallhacks.spark.util.SessionUtils;
-import net.minecraft.client.Minecraft;
-import net.minecraft.entity.Entity;
-import net.minecraft.entity.passive.EntityTameable;
-import net.minecraft.entity.passive.EntityWolf;
-import net.minecraft.network.play.server.SPacketEntityTeleport;
-import net.minecraft.util.text.TextComponentTranslation;
-import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
-
-@Module.Registration(name = "TpCoordsLogger", description = "Captures destination coordinates on teleport")
-public class TpCoordsLogger extends Module {
-    @SubscribeEvent
-    public void onServer(PacketReceiveEvent event) {
-    	if (event.getPacket() instanceof SPacketEntityTeleport) {
-    		SPacketEntityTeleport test = event.getPacket();
-    		Entity entity = Minecraft.getMinecraft().world.getEntityByID(test.getEntityId());
-    		if (entity instanceof EntityTameable) {
-    			if (((EntityTameable) entity).isTamed()) {
-    				String coordsString = String.valueOf((int)test.getX()) + "," + String.valueOf((int)test.getY()) + "," + String.valueOf((int)test.getZ());
-    				String ownerName = SessionUtils.getname(((EntityTameable) entity).getOwnerId());
-    				Spark.sendInfo("[" + entity.getName() + " teleport] Coords=" + coordsString + " Owner=" + ownerName);
-    			}
-    		}
-    	}
-    }
-
-}
